// frontend/src/api/generated/models/InvalidTag.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Galerie API
 * 本APIは、Galerieアプリケーションが画像やPDF化した漫画等のマルチメディアを提供するためのAPIである。  なお、このAPI自体はコンテンツそのものの配信を行うことを想定しておらず、このAPIを利用してコンテンツのメタデータを取得し、その中に含まれる情報を使ってコンテンツを取得することを想定している。 
 *
 * The version of the OpenAPI document: 0.4.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 無効なタグの情報
 * 
 * @export
 * @interface InvalidTag
 */
export interface InvalidTag {
    /**
     * 無効となる該当タグ名
     * @type {string}
     * @memberof InvalidTag
     */
    key: string;
    /**
     * - `INVALID_KEY`: タグ名が条件を満たしていない
     * - `UNSUPPORTED_VALUE_TYPE`: タグの型がサポートされていない
     * - `INVALID_VALUE`: 値が条件を満たしていない
     * 
     * @type {InvalidTagReasonEnum}
     * @memberof InvalidTag
     */
    reason: InvalidTagReasonEnum;
}


/**
 * @export
 */
export const InvalidTagReasonEnum = {
    InvalidKey: 'INVALID_KEY',
    UnsupportedValueType: 'UNSUPPORTED_VALUE_TYPE',
    InvalidValue: 'INVALID_VALUE'
} as const;
export type InvalidTagReasonEnum = typeof InvalidTagReasonEnum[keyof typeof InvalidTagReasonEnum];


/**
 * Check if a given object implements the InvalidTag interface.
 */
export function instanceOfInvalidTag(value: object): value is InvalidTag {
    if (!('key' in value) || value['key'] === undefined) return false;
    if (!('reason' in value) || value['reason'] === undefined) return false;
    return true;
}

export function InvalidTagFromJSON(json: any): InvalidTag {
    return InvalidTagFromJSONTyped(json, false);
}

export function InvalidTagFromJSONTyped(json: any, ignoreDiscriminator: boolean): InvalidTag {
    if (json == null) {
        return json;
    }
    return {
        
        'key': json['key'],
        'reason': json['reason'],
    };
}

export function InvalidTagToJSON(json: any): InvalidTag {
    return InvalidTagToJSONTyped(json, false);
}

export function InvalidTagToJSONTyped(value?: InvalidTag | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'key': value['key'],
        'reason': value['reason'],
    };
}